// frontend/src/services/api/models/PayslipUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PayslipUpdate = {
    employee_id?: (number | null);
    pay_period_start_date?: (string | null);
    pay_period_end_date?: (string | null);
    payment_date?: (string | null);
    total_earnings?: (number | null);
    total_deductions?: (number | null);
    net_pay?: (number | null);
    bank_account_number?: (string | null);
    notes?: (string | null);
};