// frontend/src/services/api/models/PayslipDetailCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SalaryComponentType } from './SalaryComponentType';
export type PayslipDetailCreate = {
    component_id: number;
    component_name: string;
    component_type: SalaryComponentType;
    is_earning: boolean;
    amount: number;
};