// frontend/src/services/api/models/SalaryComponentInDB.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SalaryComponentType } from './SalaryComponentType';
export type SalaryComponentInDB = {
    name: string;
    type: SalaryComponentType;
    is_earning: boolean;
    description?: (string | null);
    id: number;
};